Add all bean in the package presentation.bean
