+++ src/main/java/org/books/presentation/bean/RegisterBean.java
@@ -12,7 +12,6 @@ import javax.inject.Named;
 import org.books.application.Bookstore;
 import org.books.application.BookstoreException;
 import org.books.data.entity.Customer;
-import org.books.presentation.CustomerBean;
 import org.books.util.MessageFactory;
 
 /**
