+++ src/main/java/org/books/presentation/filter/LoginFilter.java
@@ -16,7 +16,7 @@ import javax.servlet.ServletResponse;
 import javax.servlet.annotation.WebFilter;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import org.books.presentation.CustomerBean;
+import org.books.presentation.bean.CustomerBean;
 
 /**
  *
